Moved objects into `src/utils/objects`.
