+++ src/utils.rs
@@ -237,674 +237,13 @@ pub fn isqrt(mut num: i64) -> i64 {
 /******************************************************************************
  * Objects.
  *****************************************************************************/
+mod objects;
+pub use objects::long::Long;
+pub use objects::pandigital_checker::PandigitalChecker;
+pub use objects::sieve_of_atkin::SieveOfAtkin;
+pub use objects::poker_hand::PokerHand;
+pub use objects::fraction::Fraction;
 
-/// Arbitrary-precision integer type which stores digits of a positive number
-/// in base 1_000_000_000.
-#[derive(Clone, Eq, PartialEq)]
-pub struct Long {
-    digits: Vec<u32>,
-}
-impl Long {
-    /// Construct an arbitrary-precision integer from a big-endian string of
-    /// decimal digits.
-    ///
-    /// * `s`
-    ///
-    /// -> Arbitrary-precision integer.
-    pub fn new(s: &str) -> Long {
-        let mut long = Long { digits: vec![] };
-        let mut idx = s.len();
-        loop {
-            let (lower, upper) = (std::cmp::max(idx, 9) - 9, idx);
-            long.digits.push(s[lower..upper].parse().unwrap());
-            if lower == 0 {
-                break;
-            }
-            idx = lower;
-        }
-        long
-    }
-    pub fn from(digit: u32) -> Long {
-        if digit >= 1_000_000_000 {
-            panic!("argument is too large to be a digit of this arbitrary-precision type");
-        }
-        Long { digits: vec![digit] }
-    }
-    pub fn reverse(&self) -> Long {
-        Long::new(&self.to_string().chars().rev().collect::<String>())
-    }
-    /// Calculate the factorial of a non-negative number.
-    ///
-    /// * `num` - Number whose factorial is to be calculated.
-    pub fn factorial(num: u32) -> Long {
-        if num == 0 || num == 1 {
-            return Long::from(1);
-        }
-        if num == 2 {
-            return Long::from(2);
-        }
-
-        // Multiply the extremes, converging towards the centre. For example,
-        // 9! is computed as (9 × 1) × (8 × 2) × (7 × 3) × (6 × 4) × 5, whereas
-        // 10! is computed as (10 × 1) × (9 × 2) × (8 × 3) × (7 × 4) × (6 × 5).
-        let partials = num / 2;
-        let mut result = if num % 2 == 1 {
-            Long::from(partials + 1)
-        } else {
-            Long::from(1)
-        };
-        let (mut multiplicand, mut delta) = (num, num);
-        for _ in 0..partials {
-            result *= multiplicand;
-            delta -= 2;
-            multiplicand += delta;
-        }
-        result
-    }
-    /// Obtain the number of decimal digits of this number (i.e. its length).
-    ///
-    /// -> Length.
-    pub fn len(&self) -> usize {
-        match self.digits.len() {
-            0 => 0,
-            1 if self.digits[0] == 0 => 0,
-            len => (len - 1) * 9 + self.digits.last().unwrap().to_string().len(),
-        }
-    }
-    /// Calculate the sum of all decimal digits of this number.
-    ///
-    /// -> Sum.
-    pub fn sum(&self) -> i64 {
-        self.digits
-            .iter()
-            .map(|&digit| Digits::new(digit as i64).sum::<i64>())
-            .sum()
-    }
-    /// Raise this number to the given power.
-    ///
-    /// * `exp` - Power.
-    ///
-    /// -> Value of this number raised to the given power.
-    pub fn pow(&self, mut exp: u32) -> Long {
-        // Multiplication is expensive, so these checks will improve
-        // performance.
-        let mut multiplier = Long::from(1);
-        if exp == 0 {
-            return multiplier;
-        }
-        let mut base = self.clone();
-        if exp == 1 {
-            return base;
-        }
-        loop {
-            if exp % 2 == 1 {
-                multiplier = &multiplier * &base;
-            }
-            if exp == 1 {
-                return multiplier;
-            }
-            exp /= 2;
-            base = &base * &base;
-        }
-    }
-    fn adc(a: u32, b: u32, carry: bool) -> (u32, bool) {
-        let sum = a + b + carry as u32;
-        if sum >= 1_000_000_000 {
-            (sum - 1_000_000_000, true)
-        } else {
-            (sum, false)
-        }
-    }
-    fn mlc(a: u32, b: u32, carry: u32) -> (u32, u32) {
-        let product = a as u64 * b as u64 + carry as u64;
-        ((product % 1_000_000_000) as u32, (product / 1_000_000_000) as u32)
-    }
-}
-impl std::ops::AddAssign<&Long> for Long {
-    fn add_assign(&mut self, other: &Long) {
-        self.digits
-            .resize(std::cmp::max(self.digits.len(), other.digits.len()), 0);
-        let mut carry = false;
-        for (sd, od) in self
-            .digits
-            .iter_mut()
-            .zip(other.digits.iter().chain(std::iter::repeat(&0)))
-        {
-            (*sd, carry) = Long::adc(*sd, *od, carry);
-        }
-        if carry {
-            self.digits.push(1);
-        }
-    }
-}
-impl std::ops::Add<&Long> for &Long {
-    type Output = Long;
-    fn add(self, other: &Long) -> Long {
-        let mut result = self.clone();
-        result += other;
-        result
-    }
-}
-impl std::ops::MulAssign<u32> for Long {
-    fn mul_assign(&mut self, other: u32) {
-        let mut carry = 0;
-        for sd in self.digits.iter_mut() {
-            (*sd, carry) = Long::mlc(*sd, other, carry);
-        }
-        while carry > 0 {
-            self.digits.push(carry % 1_000_000_000);
-            carry /= 1_000_000_000;
-        }
-    }
-}
-impl std::ops::Mul<u32> for &Long {
-    type Output = Long;
-    fn mul(self, other: u32) -> Long {
-        let mut result = self.clone();
-        result *= other;
-        result
-    }
-}
-impl std::ops::Mul<&Long> for &Long {
-    type Output = Long;
-    fn mul(self, other: &Long) -> Long {
-        let mut result = Long { digits: vec![] };
-        for (pad, od) in other.digits.iter().enumerate() {
-            let mut partial_product = Long { digits: vec![0; pad] };
-            let mut carry = 0;
-            for sd in self.digits.iter() {
-                let sum_carry = Long::mlc(*sd, *od, carry);
-                partial_product.digits.push(sum_carry.0);
-                carry = sum_carry.1;
-            }
-            if carry > 0 {
-                partial_product.digits.push(carry);
-            }
-            result += &partial_product;
-        }
-        result
-    }
-}
-impl std::fmt::Display for Long {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        self.digits
-            .iter()
-            .rev()
-            .enumerate()
-            .fold(Ok(()), |result, (index, digit)| {
-                result.and_then(|_| {
-                    if index == 0 {
-                        write!(f, "{}", digit)
-                    } else {
-                        write!(f, "{:0>9}", digit)
-                    }
-                })
-            })
-    }
-}
-impl std::fmt::Debug for Long {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        let result = write!(f, "digits = |");
-        self.digits
-            .iter()
-            .rev()
-            .fold(result, |result, &digit| result.and_then(|_| write!(f, "{}|", digit)))
-    }
-}
-
-/// Check whether a bunch of numbers are pandigital with respect to the given
-/// range.
-pub struct PandigitalChecker {
-    seen: [bool; 10],
-    min_digit: usize,
-    max_digit: usize,
-}
-impl PandigitalChecker {
-    pub fn new(min_digit: usize, max_digit: usize) -> PandigitalChecker {
-        PandigitalChecker {
-            seen: [false; 10],
-            min_digit,
-            max_digit,
-        }
-    }
-    pub fn renew(&mut self) {
-        self.seen = [false; 10];
-    }
-    /// Update the internal array with the digits of a given number.
-    ///
-    /// * `num` - Number to update with.
-    ///
-    /// -> Whether all digits of the number were in range and not seen earlier.
-    pub fn update(&mut self, num: i64) -> bool {
-        for digit in Digits::new(num) {
-            let digit = digit as usize;
-            if digit < self.min_digit || digit > self.max_digit || self.seen[digit] {
-                return false;
-            }
-            self.seen[digit] = true;
-        }
-        true
-    }
-    /// Check whether all digits in the range have been seen. This indicates
-    /// pandigitality only if used in tandem with the above method.
-    ///
-    /// -> Pandigitality.
-    pub fn check(&self) -> bool {
-        self.seen
-            .iter()
-            .skip(self.min_digit)
-            .take(self.max_digit - self.min_digit + 1)
-            .all(|&digit_seen| digit_seen)
-    }
-}
-
-/// A. O. L. Atkin and D. J. Bernstein, "Prime Sieves Using Binary Quadratic
-/// Forms", in Mathematics of Computation, vol. 73, no. 246, pp. 1023–1030.
-/// Generate prime numbers using the sieve of Atkin. This sieves prime numbers
-/// up to 1000000000 nearly twice as fast as my wheel-factorised sieve of
-/// Eratosthenes implementation (which I have now removed). It only determines
-/// the primality of numbers coprime to 60, because other numbers are
-/// guaranteed to be composite. (Exceptions 2, 3 and 5 are handled separately.)
-pub struct SieveOfAtkin {
-    limit: usize,
-    limit_rounded: usize,
-    limit_rounded_isqrt: usize,
-    sieve: Vec<u16>,
-}
-impl SieveOfAtkin {
-    // Consecutive differences between coprime residues modulo 60: 1, 7, 11,
-    // 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 49, 53 and 59.
-    const OFFSETS: [usize; 16] = [6, 4, 2, 4, 2, 4, 6, 2, 6, 4, 2, 4, 2, 4, 6, 2];
-    // Position of the bit indicating the primality of a coprime residue modulo
-    // 60 in a 16-element bitfield. For non-coprime residues, the value is 16.
-    const SHIFTS: [u8; 60] = [
-        16, 0, 16, 16, 16, 16, 16, 1, 16, 16, 16, 2, 16, 3, 16, 16, 16, 4, 16, 5, 16, 16, 16, 6, 16, 16, 16, 16, 16,
-        7, 16, 8, 16, 16, 16, 16, 16, 9, 16, 16, 16, 10, 16, 11, 16, 16, 16, 12, 16, 13, 16, 16, 16, 14, 16, 16, 16,
-        16, 16, 15,
-    ];
-}
-impl SieveOfAtkin {
-    /// Construct the sieve of Atkin up to and including the given number.
-    ///
-    /// * `limit` - Non-strict upper bound.
-    ///
-    /// -> Sieve of Atkin.
-    pub fn new(limit: usize) -> SieveOfAtkin {
-        // Strict upper bound divisible by 60.
-        let limit_rounded = (limit - limit % 60)
-            .checked_add(60)
-            .expect("overflow detected; argument too large");
-        let mut sieve_of_atkin = SieveOfAtkin {
-            limit,
-            limit_rounded,
-            limit_rounded_isqrt: isqrt(limit_rounded as i64) as usize,
-            sieve: vec![0; limit / 60 + 1],
-        };
-        sieve_of_atkin.init();
-        sieve_of_atkin
-    }
-    fn init(&mut self) {
-        self.algorithm_3_1(1);
-        self.algorithm_3_1(13);
-        self.algorithm_3_1(17);
-        self.algorithm_3_1(29);
-        self.algorithm_3_1(37);
-        self.algorithm_3_1(41);
-        self.algorithm_3_1(49);
-        self.algorithm_3_1(53);
-        self.algorithm_3_2(7);
-        self.algorithm_3_2(19);
-        self.algorithm_3_2(31);
-        self.algorithm_3_2(43);
-        self.algorithm_3_3(11);
-        self.algorithm_3_3(23);
-        self.algorithm_3_3(47);
-        self.algorithm_3_3(59);
-
-        // Mark composite all numbers divisible by the squares of primes.
-        let mut num: usize = 1;
-        let mut offset = SieveOfAtkin::OFFSETS.iter().cycle();
-        'sieve: for sieve_idx in 0..self.sieve.len() {
-            for shift in 0..16 {
-                if self.sieve[sieve_idx] >> shift & 1 == 1 {
-                    let num_sqr = num.pow(2);
-                    for multiple in (num_sqr..)
-                        .step_by(num_sqr)
-                        .take_while(|&multiple| multiple < self.limit_rounded)
-                    {
-                        self.sieve[multiple / 60] &= !(1u32 << SieveOfAtkin::SHIFTS[multiple % 60]) as u16;
-                    }
-                }
-                num += offset.next().unwrap();
-                if num > self.limit_rounded_isqrt {
-                    break 'sieve;
-                }
-            }
-        }
-    }
-    fn algorithm_3_1(&mut self, delta: i32) {
-        for f in 1..=15 {
-            for g in (1..=30).step_by(2) {
-                let quadratic = 4 * f * f + g * g;
-                if delta == quadratic % 60 {
-                    self.algorithm_4_1(delta, f, g, quadratic / 60);
-                }
-            }
-        }
-    }
-    fn algorithm_3_2(&mut self, delta: i32) {
-        for f in (1..=10).step_by(2) {
-            for g in [2, 4, 8, 10, 14, 16, 20, 22, 26, 28] {
-                let quadratic = 3 * f * f + g * g;
-                if delta == quadratic % 60 {
-                    self.algorithm_4_2(delta, f, g, quadratic / 60);
-                }
-            }
-        }
-    }
-    fn algorithm_3_3(&mut self, delta: i32) {
-        for (f, gstart) in (1..=10).zip([2, 1].into_iter().cycle()) {
-            for g in (gstart..=30).step_by(2) {
-                let quadratic = 3i32 * f * f - g * g;
-                // Remainder can be negative, so perform modulo operation.
-                if delta == quadratic.rem_euclid(60) {
-                    self.algorithm_4_3(delta, f, g, quadratic.div_euclid(60));
-                }
-            }
-        }
-    }
-    fn algorithm_4_1(&mut self, delta: i32, f: i32, g: i32, h: i32) {
-        let (mut x, mut y0, mut k0) = (f as i64, g as i64, h as i64);
-        while k0 < self.sieve.len() as i64 {
-            (k0, x) = (k0 + 2 * x + 15, x + 15);
-        }
-        loop {
-            (k0, x) = (k0 - 2 * x + 15, x - 15);
-            if x <= 0 {
-                return;
-            }
-            while k0 < 0 {
-                (k0, y0) = (k0 + y0 + 15, y0 + 30);
-            }
-            let (mut k, mut y) = (k0, y0);
-            while k < self.sieve.len() as i64 {
-                self.sieve[k as usize] ^= 1u16 << SieveOfAtkin::SHIFTS[delta as usize];
-                (k, y) = (k + y + 15, y + 30);
-            }
-        }
-    }
-    fn algorithm_4_2(&mut self, delta: i32, f: i32, g: i32, h: i32) {
-        let (mut x, mut y0, mut k0) = (f as i64, g as i64, h as i64);
-        while k0 < self.sieve.len() as i64 {
-            (k0, x) = (k0 + x + 5, x + 10);
-        }
-        loop {
-            (k0, x) = (k0 - x + 5, x - 10);
-            if x <= 0 {
-                return;
-            }
-            while k0 < 0 {
-                (k0, y0) = (k0 + y0 + 15, y0 + 30);
-            }
-            let (mut k, mut y) = (k0, y0);
-            while k < self.sieve.len() as i64 {
-                self.sieve[k as usize] ^= 1u16 << SieveOfAtkin::SHIFTS[delta as usize];
-                (k, y) = (k + y + 15, y + 30);
-            }
-        }
-    }
-    fn algorithm_4_3(&mut self, delta: i32, f: i32, g: i32, h: i32) {
-        let (mut x, mut y0, mut k0) = (f as i64, g as i64, h as i64);
-        loop {
-            while k0 >= self.sieve.len() as i64 {
-                if x <= y0 {
-                    return;
-                }
-                (k0, y0) = (k0 - y0 - 15, y0 + 30);
-            }
-            let (mut k, mut y) = (k0, y0);
-            while k >= 0 && y < x {
-                self.sieve[k as usize] ^= 1u16 << SieveOfAtkin::SHIFTS[delta as usize];
-                (k, y) = (k - y - 15, y + 30);
-            }
-            (k0, x) = (k0 + x + 5, x + 10);
-        }
-    }
-    pub fn is_prime(&self, num: usize) -> bool {
-        if num < 2 {
-            return false;
-        }
-        if num == 2 || num == 3 || num == 5 {
-            return true;
-        }
-        let (num_div_60, num_mod_60) = (num / 60, num % 60);
-        if num_div_60 >= self.sieve.len() {
-            panic!("queried number is out of range of the sieve")
-        }
-        self.sieve[num_div_60] & (1u32 << SieveOfAtkin::SHIFTS[num_mod_60]) as u16 != 0
-    }
-    pub fn iter(&self) -> impl Iterator<Item = i64> + '_ {
-        let mut num: usize = 1;
-        let mut offset = SieveOfAtkin::OFFSETS.iter().cycle();
-        [2, 3, 5]
-            .into_iter()
-            .chain(
-                self.sieve
-                    .iter()
-                    .flat_map(|bitfield| (0..16).map(move |shift| bitfield >> shift & 1 == 1))
-                    .filter_map(move |is_prime| {
-                        let filtered = if is_prime { Some(num) } else { None };
-                        num += offset.next().unwrap();
-                        filtered
-                    }),
-            )
-            .take_while(|&num| num <= self.limit)
-            .map(|num| num as i64)
-    }
-}
-
-/// A hand of poker.
-#[derive(Eq, PartialEq)]
-pub struct PokerHand {
-    // Pairs in which the first number is the value of a card and the second is
-    // its suit. The pairs shall be sorted in descending order of the value.
-    hand: Vec<(u8, u8)>,
-    // Used to rank this hand on an arbitrary scale. Higher is better.
-    score: i32,
-}
-impl PokerHand {
-    const HIGH_CARD: i32 = 0;
-    const ONE_PAIR: i32 = 1 << 16;
-    const TWO_PAIRS: i32 = 2 << 16;
-    const THREE_OF_A_KIND: i32 = 3 << 16;
-    const STRAIGHT: i32 = 4 << 16;
-    const FLUSH: i32 = 5 << 16;
-    const FULL_HOUSE: i32 = 6 << 16;
-    const FOUR_OF_A_KIND: i32 = 7 << 16;
-    const STRAIGHT_FLUSH: i32 = 8 << 16;
-    const ROYAL_FLUSH: i32 = 9 << 16;
-}
-impl PokerHand {
-    pub fn new(hand: &[&str]) -> PokerHand {
-        let hand = hand
-            .iter()
-            .map(|card| {
-                let card = card.as_bytes();
-                let value = match card[0] {
-                    b'T' => 10,
-                    b'J' => 11,
-                    b'Q' => 12,
-                    b'K' => 13,
-                    b'A' => 14,
-                    value => value - b'0',
-                };
-                let suit = card[1];
-                (value, suit)
-            })
-            .collect();
-        let mut poker_hand = PokerHand { hand, score: 0 };
-        poker_hand.calculate_score();
-        poker_hand
-    }
-    fn calculate_score(&mut self) {
-        // Sorting the cards by descending order of value makes analysing them
-        // easier.
-        self.hand.sort_by(|a, b| b.cmp(a));
-        let same_suit = self.hand.iter().skip(1).all(|card| card.1 == self.hand[0].1);
-        let consecutive_values = self
-            .hand
-            .iter()
-            .enumerate()
-            .skip(1)
-            .all(|(idx, card)| card.0 == self.hand[idx - 1].0 - 1);
-
-        if same_suit && consecutive_values {
-            if self.hand[0].0 == 14 {
-                self.score = PokerHand::ROYAL_FLUSH;
-            } else {
-                self.score = PokerHand::STRAIGHT_FLUSH;
-            }
-            return;
-        }
-
-        // How many times does each value appear?
-        let mut values_frequencies = [0; 15];
-        for card in &self.hand {
-            let value = card.0 as usize;
-            values_frequencies[value] += 1;
-        }
-
-        // Suppose two hands have the same rank, four of a kind: 3 clubs,
-        // 3 diamonds, 3 hearts and 3 spades, and 6 clubs, 6 diamonds, 6 hearts
-        // and 6 spades. The second hand is the winner, because the value
-        // forming four of a kind is higher. To take care of such
-        // possibilities, assign a different weightage to each rank, depending
-        // on the values forming that rank. This weightage shall be the extra
-        // score added to the rank. (If we have two pairs, assign greater
-        // weightage to the pair with the greater value.) If the scores are
-        // tied, the greatest values in each hand are compared. Hence, if the
-        // rank is formed by all cards (e.g. straight or flush), no extra score
-        // has to be added. The weightage must be chosen carefully so that it
-        // does not cause the score to overflow into the score of the next
-        // higher rank.
-        let mut score_weightage = (1, 1 << 8, 1 << 12);
-        let (mut twos, mut threes, mut fours) = (0, 0, 0);
-        let extra_score: i32 = values_frequencies
-            .iter()
-            .enumerate()
-            .map(|(value, frequency)| match frequency {
-                2 => {
-                    twos += 1;
-                    let extra_score = score_weightage.0 * value as i32;
-                    // Increase the weightage for the next pair (if any).
-                    score_weightage.0 <<= 4;
-                    extra_score
-                }
-                3 => {
-                    threes += 1;
-                    score_weightage.1 * value as i32
-                }
-                4 => {
-                    fours += 1;
-                    score_weightage.2 * value as i32
-                }
-                _ => 0,
-            })
-            .sum();
-
-        if fours == 1 {
-            self.score = PokerHand::FOUR_OF_A_KIND + extra_score;
-            return;
-        }
-        if twos == 1 && threes == 1 {
-            self.score = PokerHand::FULL_HOUSE + extra_score;
-            return;
-        }
-        if same_suit {
-            self.score = PokerHand::FLUSH;
-            return;
-        }
-        if consecutive_values {
-            self.score = PokerHand::STRAIGHT;
-            return;
-        }
-        if threes == 1 {
-            self.score = PokerHand::THREE_OF_A_KIND + extra_score;
-            return;
-        }
-        if twos == 2 {
-            self.score = PokerHand::TWO_PAIRS + extra_score;
-            return;
-        }
-        if twos == 1 {
-            self.score = PokerHand::ONE_PAIR + extra_score;
-            return;
-        }
-        self.score = PokerHand::HIGH_CARD;
-    }
-}
-impl Ord for PokerHand {
-    fn cmp(&self, other: &PokerHand) -> std::cmp::Ordering {
-        let compare = self.score.cmp(&other.score);
-        if compare != std::cmp::Ordering::Equal {
-            return compare;
-        }
-        for (scard, ocard) in self.hand.iter().zip(other.hand.iter()) {
-            let compare = scard.0.cmp(&ocard.0);
-            if compare != std::cmp::Ordering::Equal {
-                return compare;
-            }
-        }
-        std::cmp::Ordering::Equal
-    }
-}
-impl PartialOrd for PokerHand {
-    fn partial_cmp(&self, other: &PokerHand) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-/// Rational numbers.
-#[derive(Clone)]
-pub struct Fraction {
-    numerator: Long,
-    denominator: Long,
-}
-impl Fraction {
-    pub fn from(numerator: u32, denominator: u32) -> Fraction {
-        Fraction {
-            numerator: Long::from(numerator),
-            denominator: Long::from(denominator),
-        }
-    }
-    pub fn invert(&mut self) {
-        std::mem::swap(&mut self.numerator, &mut self.denominator);
-    }
-    pub fn len(&self) -> (usize, usize) {
-        (self.numerator.len(), self.denominator.len())
-    }
-    pub fn sum(&self) -> (i64, i64) {
-        (self.numerator.sum(), self.denominator.sum())
-    }
-}
-impl std::ops::AddAssign<u32> for Fraction {
-    fn add_assign(&mut self, other: u32) {
-        self.numerator += &(&self.denominator * other);
-    }
-}
-impl std::ops::Add<u32> for &Fraction {
-    type Output = Fraction;
-    fn add(self, other: u32) -> Fraction {
-        let mut result = self.clone();
-        result += other;
-        result
-    }
-}
-impl std::fmt::Display for Fraction {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}/{}", self.numerator, self.denominator)
-    }
-}
 
 /******************************************************************************
  * Iterators.
